+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# cdata 1.0.8 2019/04/02
+
+ * Move to wrapr draw_framec().
+
 # cdata 1.0.8 2019/03/30
 
  * More column collision checks.